refactor: change return type of getId in IssueWrapper from int to long for consistency

+++ lambda/src/main/java/com/lambda/models/IssueWrapper.java
@@ -26,8 +26,8 @@ public class IssueWrapper {
         this.newStatusCode = newStatusCode;
     }
 
-    public int getId() {
-        return (int) rawIssue.getId();
+    public long getId() {
+        return rawIssue.getId();
     }
 
     public long getProjectId() {
